Only accept messages from the guild id specified in our config

+++ discord_gambler/commands/cleanse.py
@@ -1,6 +1,6 @@
 from discord.ext import commands
 from decouple import config
-from discord_gambler import _coinflip_channel
+from discord_gambler import _coinflip_channel, _guild_id
 import os
 
 
@@ -10,7 +10,7 @@ class CleanseCommand(commands.Cog):
 
     @commands.command(name="cleanse")
     async def on_cleanse_command(self, ctx):
-        if ctx.channel.name == _coinflip_channel:
+        if ctx.channel.name == _coinflip_channel and ctx.guild.id == _guild_id:
             print("Cleaning up coinflip channel...")
             for message in await ctx.channel.history(limit=100).flatten():
                 await message.delete()
